backend/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
google-generativeai==0.7.2
python-dotenv==1.0.1
Flask-Cors==4.0.1
gunicorn==22.0.0
SQLAlchemy==2.0.31
alembic==1.13.2
passlib==1.7.4
PyJWT==2.9.0
stripe==8.3.0
zeep==4.2.1
requests==2.31.0
beautifulsoup4==4.12.2
